requirements.txt
line-bot-sdk==1.16.0